Export function to get Jest configuration
