+++ packages/components/jest.config.mjs
@@ -1,6 +1,6 @@
 // ESLint thinks @terra-ui-packages/test-utils should be listed in dependencies instead of devDependencies.
 // TODO: Fix this in ESLint configuration.
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { baseConfig } from '@terra-ui-packages/test-utils';
+import { getJestConfig } from '@terra-ui-packages/test-utils';
 
-export default baseConfig;
+export default getJestConfig();
