+++ packages/core-utils/jest.config.mjs
@@ -1,10 +1,10 @@
 // ESLint thinks @terra-ui-packages/test-utils should be listed in dependencies instead of devDependencies.
 // TODO: Fix this in ESLint configuration.
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { baseConfig } from '@terra-ui-packages/test-utils';
+import { getJestConfig } from '@terra-ui-packages/test-utils';
 
 export default {
-  ...baseConfig,
+  ...getJestConfig(),
   // Ignore type tests when calculating coverage.
   coveragePathIgnorePatterns: ['\\.types\\.ts$', '\\.errors\\.ts$'],
 };
